pyproject.toml
[project]
name = "morning-market-agent"
version = "0.1.0"
description = "Generate a Morning Market Brief"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.25",
  "pydantic>=2.5",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[tool.pytest.ini_options]
pythonpath = ["."]
